src/*, tests: separate test modules into files for codecov and readability
